feat(query): separate query from observable
BREAKING CHANGES: query is now a own package
