+++ node16.Dockerfile
@@ -5,4 +5,4 @@ RUN apk add --update --no-cache python3 make g++
 COPY . .
 
 RUN yarn install --frozen-lockfile
-RUN npx lerna exec 'yarn prepack'
+RUN npx lerna exec 'npm run prepack --if-present'
