+++ packages/module-app/package.json
@@ -34,6 +34,6 @@
         "typescript": "^4.8.4"
     },
     "peerDependencies": {
-        "@equinor/fusion-observable": "^1.2.1"
+        "@equinor/fusion-query": "^0.0.0"
     }
 }
