+++ packages/module-app/tsconfig.json
@@ -11,7 +11,7 @@
       "path": "../module"
     },
     {
-      "path": "../../utils/observable"
+      "path": "../../utils/query"
     },
     {
       "path": "../module-http"
