+++ packages/module-context/package.json
@@ -30,16 +30,17 @@
         "url": "git+https://github.com/equinor/fusion-framework.git",
         "directory": "packages/modules-context"
     },
+    "dependencies": {
+        "@equinor/fusion-query": "^0.0.0"
+    },
     "devDependencies": {
         "@equinor/fusion-framework-module": "^1.2.9",
         "@equinor/fusion-framework-module-event": "^1.1.5",
         "@equinor/fusion-framework-module-services": "^0.5.9",
-        "@equinor/fusion-observable": "^1.5.1",
         "rxjs": "^7.5.7"
     },
     "peerDependencies": {
         "@equinor/fusion-framework-module": "^1.2.5",
-        "@equinor/fusion-observable": "^1.2.1",
         "rxjs": "^7.5.7"
     }
 }
