+++ packages/react-module-context/src/useQueryContext.ts
@@ -4,7 +4,7 @@ import {
     IContextProvider,
 } from '@equinor/fusion-framework-module-context';
 import { useModule } from '@equinor/fusion-framework-react-module';
-import { useDebounceQuery } from '@equinor/fusion-observable/react';
+import { useDebounceQuery } from '@equinor/fusion-query/react';
 
 export const useQueryContext = (provider: IContextProvider, options?: { debounce?: number }) => {
     const args = Object.assign({}, { debounce: 500 }, options);
