+++ packages/react-module-context/tsconfig.json
@@ -9,7 +9,7 @@
   },
   "references": [
     {
-      "path": "../../utils/observable"
+      "path": "../../utils/query"
     },
     {
       "path": "../module-context"
