+++ packages/test-app/package.json
@@ -25,6 +25,7 @@
         "@equinor/fusion-framework-react": "^1.3.19",
         "@equinor/fusion-framework-react-app": "^1.3.6",
         "@equinor/fusion-observable": "^1.5.1",
+        "@equinor/fusion-query": "^0.0.0",
         "@equinor/fusion-react-progress-indicator": "^0.1.5",
         "@equinor/fusion-react-styles": "^0.5.5",
         "@tanstack/react-query": "^4.2.3",
