+++ packages/test-app/tsconfig.json
@@ -31,6 +31,9 @@
       "path": "../../utils/observable"
     },
     {
+      "path": "../../utils/query"
+    },
+    {
       "path": "../react-app"
     },
     {
