+++ utils/observable/src/react/index.ts
@@ -7,5 +7,3 @@ export * from './useObservableSelector';
 export * from './useObservableSelectorState';
 export * from './useObservableSubscription';
 export * from './useObservableInput';
-
-export * from './useDebounceQuery';
