+++ utils/query/src/client/QueryClientError.ts
